+++ data/index.js
@@ -1,6 +1,10 @@
 const makeChainFeeRateResponse = require('./make_chain_fee_rate_response');
 const makeForwardsResponse = require('./make_forwards_response');
 const makeFundPsbtResponse = require('./make_fund_psbt_response');
+const makeGetChainTxResponse = require('./make_get_chain_tx_response');
+const makeGetChannelsResponse = require('./make_get_channels_response');
+const makeGetPendingChanResponse = require('./make_get_pending_chan_response');
+const makeGetUtxosResponse = require('./make_get_utxos_response');
 const makeInvoice = require('./make_invoice');
 const makePayViaRoutesResponse = require('./make_pay_via_routes_response');
 const makePaymentData = require('./make_payment_data');
@@ -14,6 +18,10 @@ module.exports = {
   makeChainFeeRateResponse,
   makeForwardsResponse,
   makeFundPsbtResponse,
+  makeGetChainTxResponse,
+  makeGetChannelsResponse,
+  makeGetPendingChanResponse,
+  makeGetUtxosResponse,
   makeInvoice,
   makePayViaRoutesResponse,
   makePaymentData,