+++ data/make_get_pending_chan_response.js
@@ -1,0 +1,14 @@
+/** Make a response for pending channels
+
+  {}
+
+  @returns
+  <PendingChannels Response Object>
+*/
+module.exports = ({}) => {
+  return {
+    pending_force_closing_channels: [],
+    pending_open_channels: [],
+    waiting_close_channels: [],
+  };
+};